requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
idna==2.6
lxml==4.1.1
pymongo==3.5.1
python-slugify==1.2.4
requests==2.18.4
Unidecode==0.4.21
urllib3==1.22
wheel==0.24.0
